Add speedup categories in qualification summary output
