+++ user_tools/src/spark_rapids_pytools/resources/qualification-conf.yaml
@@ -27,7 +27,20 @@ toolOutput:
         - Unsupported Operators Stage Duration Percent
         - Speed Up Estimation Model
       mapColumns:
-        Recommendation: 'Speedup Based Recommendation'
+        - srcCol: 'Recommendation'
+          dstCol: 'Speedup Based Recommendation'
+        - srcCol: 'Estimated GPU Speedup'
+          dstCol: 'Estimated GPU Speedup Category'
+          recommendationRanges:
+            - title: 'Small'
+              lowerBound: 1.3
+              upperBound: 2.0
+            - title: 'Medium'
+              lowerBound: 2.0
+              upperBound: 3.0
+            - title: 'Large'
+              lowerBound: 3.0
+              upperBound: 1000000.0
       recommendations:
         speedUp:
           columnName: 'Speedup Based Recommendation'
@@ -216,9 +229,7 @@ local:
       outputColumns:
         - 'App ID'
         - 'App Name'
-        - 'App Duration'
-        - 'Estimated GPU Speedup'
-        - 'Unsupported Operators Stage Duration Percent'
+        - 'Estimated GPU Speedup Category'
       sortingColumns:
         - 'Estimated GPU Speedup'
         - 'Unsupported Operators Stage Duration Percent'
@@ -233,23 +244,6 @@ local:
           - columnName: 'Unsupported Operators Stage Duration Percent'
             lowerBound: 0.0
             upperBound: 25.0
-      output: # Configs related to output
-        columns:
-          - 'App ID'
-          - 'App Name'
-          - 'Estimated GPU Speedup'
-        remap:
-          - columnName: 'Estimated GPU Speedup'
-            recommendationRanges:
-              - title: 'Small'
-                lowerBound: 1.3
-                upperBound: 2.0
-              - title: 'Medium'
-                lowerBound: 2.0
-                upperBound: 3.0
-              - title: 'Large'
-                lowerBound: 3.0
-                upperBound: 1000000.0
     predictionModel:
       outputDirectory: 'xgboost_predictions'
       files:
